+++ routes/_app.tsx
@@ -1,9 +1,21 @@
 import { AppProps } from "$fresh/server.ts";
+import { Head } from "$fresh/runtime.ts";
 
 export default function App({ Component }: AppProps) {
   return (
     <>
-      <Component />
+      <Head>
+        <title>remember.land</title>
+        <link
+          rel="stylesheet"
+          href="https://cdn.jsdelivr.net/gh/CarcajadaArtificial/lunchbox@0.1.1/static/style.css"
+        />
+      </Head>
+      <body class="clr-bg-panel clr-txt-base txt-paragraph">
+        <div class="_screen">
+          <Component />
+        </div>
+      </body>
     </>
   );
 }